0329/pyproject.toml
[project]
name = "0329"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.0",
    "django>=5.2.12",
    "djangorestframework>=3.17.1",
    "path>=17.1.1",
    "pymysql>=1.1.1",
]
